+++ routes/authentication.js
@@ -17,6 +17,7 @@ const { userModel }         = require('../schema');
 // ==============================
 // middleware
 const { _pre, _end }        = require('../config/middleware');
+router.all('*', _pre.usePassport);
 
 
 
@@ -31,7 +32,7 @@ router
         if (req.isAuthenticated()) return res.redirect('/console/dashboard');
         res.render('./console/account/signup')
     })
-    .post(_pre.passwordValidation, _pre.usePassport, _end.wrapAsync(async (req, res) => {
+    .post(_pre.passwordValidation, _end.wrapAsync(async (req, res) => {
         const newUser = await userModel.register(new userModel(req.body), req.body.password.new);
         req.logIn(newUser, err => {
             if (err) throw err;
@@ -45,13 +46,13 @@ router
 router
     .route('/signin')
     .all(_pre.prependTitleTag('Sign In'))
-    .get(_pre.usePassport, (req, res) => {
+    .get((req, res) => {
         if (req.isAuthenticated()) return res.redirect('/console');
         res.render('./console/account/signin');
     })
-    .patch(_pre.usePassport, (req, res, next) => {
+    .patch((req, res, next) => {
         if (req.isAuthenticated()) return res.redirect('/console');
-        passport.authenticate('local', (err, authUser) => {   // note: async/await will handle error; otherwise handle by calling next();
+        passport.authenticate('local', (err, authUser) => {
             // exception
             if (err) return next(err);
             if (!authUser) return next(new Error('Wrong email/username or password!'));
@@ -60,28 +61,30 @@ router
             req.logIn(authUser, err => {
                 if (err) return next(err);
 
-                // session population
-                req.session.cookie.expires = (req.body.isPersisted) ? new Date(Date.now() + 14 * 24 * 3600 * 1000) : false;
+                // session
+                req.session.cookie.expires = (req.body.isPersisted)
+                    ? new Date(Date.now() + 14 * 24 * 3600 * 1000)
+                    : false;
                 req.session.user = {
                     _id: authUser._id,
-                    nickname: authUser.nickname,
                     picture: authUser.picture,
+                    nickname: authUser.nickname,
                 };
 
-                // other response
+                // client response
+                req.flash('info', `Welcome back ${authUser.username}`);
                 const returnTo = req.session.returnTo;
                 delete req.session.returnTo;
-                req.flash('info', `Welcome back ${authUser.username}`);
                 return res.redirect(returnTo || '/console/dashboard');
             });
-        })(req, res)
+        })(req, res);
     });
 
 
 // sign-out
 router
     .route('/signout')
-    .get(_pre.usePassport, (req, res) => {
+    .get((req, res) => {
         if (req.isAuthenticated()) {
             req.logout();
             req.flash('pass', true);