+++ HISTORY.md
@@ -1,3 +1,12 @@
+### 0.2.5 / 2017-11-14
+> App
+- Fixed the sign-out flash message.
+- Rearranged 'authentication' route.
+
+> UI
+- n/a
+
+
 ### 0.2.4 / 2017-11-14
 > App
 - Removed the Mongo's anti-pattern as in `docList` schema.